src/image_translator/segmenter/requirements.txt
fastapi
pillow
numpy
wtpsplit[onnx-gpu]
uvicorn
python-multipart